Really deduplicate seller info requests
Seller info URLs are different even for the same seller.
Previously I assumed that seller URLs were the same for the same seller,
and that deduplicating theses URLs would be enough.
That was incorrect.
We don't want multiple requests for basically the same info
as it triggers Amazon's HTTP503 request throttling.
This patch extracts the seller IDs and constructs same URLs for the same sellers.